[TWINFACES-458] feat: resolve PR comments
