+++ src/entities/link/api/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-fetch-by-id";
+export * from "./use-search-v1";
