+++ src/entities/link/api/types.ts
@@ -1,4 +1,4 @@
-import { components } from "@/shared/api/generated/schema";
+import { components, operations } from "@/shared/api/generated/schema";
 import { RequireFields } from "@/shared/libs";
 
 export type TwinLinkView = components["schemas"]["TwinLinkViewV1"];
@@ -32,5 +32,8 @@ export type LinkFilterKeys =
 export type LinkFilters = Partial<
   Pick<components["schemas"]["LinkSearchRqV1"], LinkFilterKeys>
 >;
+
+export type UpdateLinkRequestBody = components["schemas"]["LinkUpdateV1"];
+export type QueryLinkViewV1 = operations["linkViewV1"]["parameters"]["query"];
 
 export type CreateLinkRequestBody = components["schemas"]["LinkCreateV1"];
