+++ src/entities/link/libs/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from "./use-search-v1";
 export * from "./use-filters";
 export * from "./use-select-adapter";
