+++ src/entities/twin-class-link/api/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from "./use-search-v1";
-export * from "./use-fetch-by-id";
+export * from "../../../link/api/hooks/use-fetch-by-id";
 export * from "./use-update-link";
