+++ src/entities/twin-class-link/api/index.ts
@@ -1,3 +1,2 @@
 export * from "./api-service";
-export * from "./types";
 export * from "./hooks";
